Start re-implementing the OOP API in lua. Add lua scripting console.

+++ src/components/warpdrive.h
@@ -11,12 +11,12 @@ public:
     constexpr static float heat_per_warp = 0.02f;
 
     // Config
-    int8_t max_level = 4;
+    int max_level = 4;
     float speed_per_level = 1000;
     float energy_warp_per_second = 1.7f;
 
     // Runtime
-    int8_t request = 0; // [input] Level of warp requested, from 0 to max_level
+    int request = 0; // [input] Level of warp requested, from 0 to max_level
     float current = 0.0f; // [output] Current active warp amount, from 0.0 to 4.0
 };
